# contracts/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ТехноСпортФест 2026 API",
    "version": "1.0.0",
    "description": "Канонический контракт MVP. Времена UTC. JSON camelCase. Неизвестные поля запросов отклоняются. Списки сортируются детерминированно: competitions по startsAt,id; participants по fullName,registrationId; results по place,fullName,id; news/documents по publishedAt DESC,id DESC; rating по points DESC,fullName,id."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
...
  "paths": {
...
    "/competitions/{competitionId}/results/{registrationId}": {
      "put": {
        "tags": [
          "Results"
        ],
        "summary": "Создать или заменить черновик результата",
        "operationId": "upsertResultDraft",
        "x-owner": "B2",
        "x-milestone": "v1",
        "responses": {
          "200": {
            "description": "Успех",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultDraft"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "422": {
            "$ref": "#/components/responses/422"
          }
        },
        "description": "Только organizer. Заявка должна принадлежать этому соревнованию. Допустимо только status=published; повторный PUT до публикации заменяет черновик.",
        "security": [
          {
            "accessCookie": []
          }
        ],
        "parameters": [
          {
            "name": "competitionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "registrationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResultDraftRequest"
              }
            }
          }
        },
        "x-origin-check": true
      }
    },
    "/competitions/{competitionId}/results/drafts": {
      "get": {
        "tags": [
          "Results"
        ],
        "summary": "Черновики результатов",
        "operationId": "listResultDrafts",
        "x-owner": "B2",
        "x-milestone": "v1",
        "responses": {
          "200": {
            "description": "Успех",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultDraftPage"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        },
        "description": "Только organizer. Не включает опубликованные результаты.",
        "security": [
          {
            "accessCookie": []
          }
        ],
        "parameters": [
          {
            "name": "competitionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          }
        ]
      }
    },
    "/competitions/{competitionId}/results/publish": {
      "post": {
        "tags": [
          "Results"
        ],
        "summary": "Опубликовать все черновики",
        "operationId": "publishResults",
        "x-owner": "B2",
        "x-milestone": "v1",
        "responses": {
          "200": {
            "description": "Успех",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishResultsResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "409": {
            "$ref": "#/components/responses/409"
          }
        },
        "description": "Только organizer. Требуется >=1 черновик. Одна транзакция публикует все черновики и завершает соревнование. Повторный вызов — 409 ALREADY_COMPLETED.",
        "security": [
          {
            "accessCookie": []
          }
        ],
        "parameters": [
          {
            "name": "competitionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "x-origin-check": true
      }
    },
    "/competitions/{competitionId}/results": {
      "get": {
        "tags": [
          "Results"
        ],
        "summary": "Публичные результаты соревнования",
        "operationId": "listCompetitionResults",
        "x-owner": "B2",
        "x-milestone": "v1",
        "responses": {
          "200": {
            "description": "Успех",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultPage"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        },
        "description": "Только опубликованные результаты; draft соревнование для публики даёт 404.",
        "security": [],
        "parameters": [
          {
            "name": "competitionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          }
        ]
      }
    },
    "/me/results": {
      "get": {
        "tags": [
          "Results"
        ],
        "summary": "Мои опубликованные результаты",
        "operationId": "listMyResults",
        "x-owner": "B2",
        "x-milestone": "v2",
        "responses": {
          "200": {
            "description": "Успех",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultPage"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          }
        },
        "security": [
          {
            "accessCookie": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          }
        ]
      }
    },
    "/ratings": {
      "get": {
        "tags": [
          "Rating"
        ],
        "summary": "Общий рейтинг",
        "operationId": "getRatings",
        "x-owner": "B2",
        "x-milestone": "v1",
        "responses": {
          "200": {
            "description": "Успех",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Rating"
                }
              }
            }
          },
          "422": {
            "$ref": "#/components/responses/422"
          }
        },
        "description": "Только опубликованные результаты. rank = 1 + число спортсменов с большим числом баллов; ничьи делят место. Пустой рейтинг — items=[],total=0.",
        "security": [],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "disciplineId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      }
    },
    "/me/rating": {
      "get": {
        "tags": [
          "Rating"
        ],
        "summary": "Моя позиция",
        "operationId": "getMyRating",
        "x-owner": "B2",
        "x-milestone": "v2",
        "responses": {
          "200": {
            "description": "Успех",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyRating"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "422": {
            "$ref": "#/components/responses/422"
          }
        },
        "description": "Если результатов нет, points=0, resultsCount=0, rank=null.",
        "security": [
          {
            "accessCookie": []
          }
        ],
        "parameters": [
          {
            "name": "disciplineId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "accessCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "tsf_access",
        "description": "JWT, HttpOnly; SameSite=Lax; Secure в production; срок 24 часа"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable machine-readable error code"
          },
          "message": {
            "type": "string",
            "description": "Human-readable Russian message"
          },
          "fields": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "description": "Optional field errors"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
...
      "ResultDraftRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "place": {
            "type": "integer",
            "minimum": 1
          },
          "scoreText": {
            "type": "string",
            "maxLength": 500,
            "nullable": true
          }
        },
        "required": [
          "place"
        ]
      },
      "ResultDraft": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "registrationId": {
            "type": "string",
            "format": "uuid"
          },
          "competitionId": {
            "type": "string",
            "format": "uuid"
          },
          "athleteId": {
            "type": "string",
            "format": "uuid"
          },
          "fullName": {
            "type": "string"
          },
          "place": {
            "type": "integer",
            "minimum": 1
          },
          "scoreText": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "draft"
            ]
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "registrationId",
          "competitionId",
          "athleteId",
          "fullName",
          "place",
          "scoreText",
          "status",
          "updatedAt"
        ]
      },
      "ResultDraftPage": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ResultDraft"
            }
          },
          "page": {
            "type": "integer",
            "minimum": 1
          },
          "pageSize": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100
          },
          "total": {
            "type": "integer",
            "minimum": 0
          }
        },
        "required": [
          "items",
          "page",
          "pageSize",
          "total"
        ]
      },
      "Result": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "registrationId": {
            "type": "string",
            "format": "uuid"
          },
          "competitionId": {
            "type": "string",
            "format": "uuid"
          },
          "athleteId": {
            "type": "string",
            "format": "uuid"
          },
          "fullName": {
            "type": "string"
          },
          "place": {
            "type": "integer",
            "minimum": 1
          },
          "scoreText": {
            "type": "string",
            "nullable": true
          },
          "points": {
            "type": "integer",
            "enum": [
              100,
              70,
              50,
              20
            ]
          },
          "publishedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "registrationId",
          "competitionId",
          "athleteId",
          "fullName",
          "place",
          "scoreText",
          "points",
          "publishedAt"
        ]
      },
      "ResultPage": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Result"
            }
          },
          "page": {
            "type": "integer",
            "minimum": 1
          },
          "pageSize": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100
          },
          "total": {
            "type": "integer",
            "minimum": 0
          }
        },
        "required": [
          "items",
          "page",
          "pageSize",
          "total"
        ]
      },
      "PublishResultsResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "competitionId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string",
            "enum": [
              "completed"
            ]
          },
          "publishedCount": {
            "type": "integer",
            "minimum": 1
          },
          "publishedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "competitionId",
          "status",
          "publishedCount",
          "publishedAt"
        ]
      },
      "RatingRow": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "athleteId": {
            "type": "string",
            "format": "uuid"
          },
          "fullName": {
            "type": "string"
          },
          "points": {
            "type": "integer",
            "minimum": 0
          },
          "rank": {
            "type": "integer",
            "minimum": 1
          },
          "resultsCount": {
            "type": "integer",
            "minimum": 1
          }
        },
        "required": [
          "athleteId",
          "fullName",
          "points",
          "rank",
          "resultsCount"
        ]
      },
      "Rating": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "disciplineId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "formula": {
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RatingRow"
            }
          },
          "page": {
            "type": "integer",
            "minimum": 1
          },
          "pageSize": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100
          },
          "total": {
            "type": "integer",
            "minimum": 0
          }
        },
        "required": [
          "disciplineId",
          "formula",
          "items",
          "page",
          "pageSize",
          "total"
        ]
      },
      "MyRating": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "athleteId": {
            "type": "string",
            "format": "uuid"
          },
          "disciplineId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "points": {
            "type": "integer",
            "minimum": 0
          },
          "rank": {
            "type": "integer",
            "minimum": 1,
            "nullable": true
          },
          "resultsCount": {
            "type": "integer",
            "minimum": 0
          }
        },
        "required": [
          "athleteId",
          "disciplineId",
          "points",
          "rank",
          "resultsCount"
        ]
      },
...
    },
    "responses": {
...
      "401": {
        "description": "Нет действующего JWT",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "code": "UNAUTHENTICATED",
              "message": "Нет действующего JWT"
            }
          }
        }
      },
      "403": {
        "description": "Недостаточно прав или неверный/отсутствующий Origin",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "code": "FORBIDDEN",
              "message": "Недостаточно прав или неверный/отсутствующий Origin"
            }
          }
        }
      },
      "404": {
        "description": "Ресурс не найден",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "code": "NOT_FOUND",
              "message": "Ресурс не найден"
            }
          }
        }
      },
      "409": {
        "description": "Конфликт состояния или уникальности",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "code": "CONFLICT",
              "message": "Конфликт состояния или уникальности"
            }
          }
        }
      },
      "422": {
        "description": "Ошибка полей",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "code": "VALIDATION_ERROR",
              "message": "Ошибка полей"
            }
          }
        }
      }
    }
  }
}